move search function into separate file
